Refactor plot building, add legend positioning, improve data padding
This commit refactors some areas of the plot building logic and improves some existing functionalities.

PlotBuilderFluent_OfType.cs was updated to improve the descriptive nature of method and property comments.
In PlotBuilderFluent_Configuration.cs, a new feature was added to position the plot legend outside the top right corner of the plot. Also, the data padding function was improved to accommodate the number of channels and the percent increment for each channel.

An additional validation was created to prevent setting data padding for zero or negative number of channels. The use of PlotLimitCalculator class was optimized in terms of memory allocation. Corresponding changes and improvements were also carried out in their respective interface classes to maintain consistency.

Lastly, a new exception message was also added on Message.cs for the new validation in the data padding method. These changes further enhance the plot building process and provide flexibility to users while creating plots.