Remove value range constraint from PlotBuilderFluent_Configuration
This change removes the limits imposed on the values of 'valueX' and 'valueY' in PlotBuilderFluent_Configuration.cs file. These constraints were preventing the usage of the plotting API with values outside the 0 to 1 range, which was unnecessarily limiting. Now, users can freely use any range of values for 'valueX' and 'valueY', enhancing the flexibility of the API.