Add color parameter to AddText method
In IPlotBuilderFluent_CanvasConfiguration, the AddText method has been adjusted to include a color parameter. This change allows users to specify the color of the annotation while adding text to the plot. Corresponding changes have been made in PlotBuilderFluent_Canvas and the documentation has also been updated to reflect this new feature.