Add cancellation tokens to plot save methods
Refactored the save methods in PlotBuilderFluent classes and BitmapProcessor to use optional cancellation tokens for the async state machine. This update provides the ability to cancel ongoing operations. In addition, a new 'CancelAllOperations' method has been added to IPlotBuilderFluentProduct and IPlotBuilderFluentCanvasProduct interfaces to cancel all operations and dispose the cancellation token source.