Add SetSourcePath method and refactor TrySave methods
Added a method SetSourcePath to the IPlotBuilderFluentCanvasConfiguration interface and removed it from IPlotBuilderFluentCanvas. Replaced all instances of the TrySaveAsync method with the TrySaveAtBmpParserAsync method in IPlotBuilderFluent_CanvasProduct and PlotBuilderFluent_CanvasProduct. BitmapProcessor was renamed to BitmapParser and all its references were updated. Files were also renamed in the simple-plotting project for consistency.